requirements.txt
pytesseract==0.3.10
opencv-python==4.8.0.74
numpy==1.24.3
Pillow==9.5.0
mss==9.0.1
pywin32==306
requests==2.31.0
python-dotenv==1.0.0
textblob==0.17.1
newsapi-python==0.2.6
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
python-multipart==0.0.6
